data_collection/pyproject.toml
[tool.poetry]
authors = ["Oliwier Szypczyn <oli.szypczyn@gmail.com>"]
description = ""
name = "src"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
feedparser = "^6.0.11"
kafka-python-ng = "^2.2.3"
python = "^3.13"
python-decouple = "^3.8"
requests = "^2.32.3"
six = "^1.16.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.pylint]
disable = [
  "E0401",
  "W0135",
  "C0301",
]

[tool.black]
line-length = 100

[tool.poetry.scripts]
data-collection = "src.app:main"
